+++ CHANGELOG.md
@@ -6,10 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 ### Added
+#### CpxAp
 - Read the global diagnosis state from the cpx-ap system with `CpxAp.read_global_diagnosis_state()`
 - Read the count of active diagnosis from the cpx-ap system with `CpxAp.read_active_diagnosis_count()`
 - Read the module index with the latest diagnosis from the cpx-ap system with `CpxAp.read_latest_diagnosis_index()`
 - Read the latest diagnosis code from the cpx-ap system with `CpxAp.read_latest_diagnosis_code()`
+#### ApModule
+- Read the diagnosis code from the module with `ApModule.read_diagnosis_code()`
+- Read the diagnosis information from the module with `ApModule.read_diagnosis_information()`
 
 ## v0.4.2 - 29.05.24
 ### Changed